Catch attempting to insert duplicate records and return an error message
This isn't the nicest implementation of this: the error message could be
a lot more informational. For one thing, it should say which value
(email or phone number) conflicted. It might also want to provide the id
of the conflicting record. In lieu of any actual requirements, however,
this will do for now.

fixes #2